update cat_facts web-nextjs with latest idioms
