+++ examples/cat_facts/web-nextjs/src/app/http.ts
@@ -1,0 +1,29 @@
+import type { HttpRequest } from "shared_types/types/shared_types";
+import { HttpResponse, HttpHeader } from "shared_types/types/shared_types";
+
+export async function request({
+  url,
+  method,
+  headers,
+}: HttpRequest): Promise<HttpResponse> {
+  const request = new Request(url, {
+    method,
+    headers: headers.map((header) => [header.name, header.value]),
+  });
+
+  const response = await fetch(request);
+
+  const responseHeaders = Array.from(
+    response.headers.entries(),
+    ([name, value]) => new HttpHeader(name, value)
+  );
+
+  const body = await response.arrayBuffer();
+
+  const httpResponse = new HttpResponse(
+    response.status,
+    responseHeaders,
+    Array.from(new Uint8Array(body))
+  );
+  return httpResponse;
+}
